requirements.txt
faiss_cpu==1.8.0.post1
langchain==0.3.7
langchain_core==0.3.15
langchain_ollama==0.2.0
numpy==1.24.4
pandas==2.0.3
pymongo==3.12.0
python-dotenv==1.0.1
streamlit==1.40.0
streamlit_chat==0.1.1
torch==2.4.1
